image-dataset-cleaner/requirements.txt
certifi==2021.10.8
fire==0.4.0
Pillow==9.1.0
PyYAML==6.0
six==1.16.0
termcolor==1.1.0
wincertstore==0.2
